services/auth/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
elasticsearch==8.11.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
celery==5.3.4
aioredis==2.0.1
httpx==0.25.2
python-dotenv==1.0.0
prometheus-client==0.19.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
isort==5.13.2
flake8==6.1.0
PyJWT==2.8.0
bcrypt==3.2.2
twilio==8.10.0